+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Julia docstrings for all public exports (module, types, attention kernels, normalization).
+
 ### Changed
 
 - Added dual MIT/Apache-2.0 licensing.